+++ sdk/python/pulumi_okta/auth/__init__.py
@@ -8,4 +8,5 @@ from .server import *
 from .server_claim import *
 from .server_policy import *
 from .server_policy_claim import *
+from .server_policy_rule import *
 from .server_scope import *